requirements.txt
amqp==2.5.2
celery==4.3.0
Flask==1.1.1
importlib-metadata==0.20
kombu==4.6.4
mock==3.0.5
more-itertools==7.2.0
Pillow==6.2.0
redis==3.3.8
six==1.12.0
Werkzeug==0.15.6
